[tests] move BCNode to separate module
